+++ app/Http/Controllers/Api/DetalleIngresoController.php
@@ -15,26 +15,50 @@ public function index()
 
     public function store(Request $request)
     {
+            $request->validate([
+            'ingreso_id' => 'required|integer|exists:ingresos,id',
+            'producto_id' => 'required|integer|exists:productos,id',
+            'cantidad' => 'required|integer|min:1',
+            'precio_compra' => 'required|numeric',
+        ]);
+
         $detalle = DetalleIngreso::create($request->all());
         return response()->json($detalle, 201);
     }
 
     public function show(string $id)
     {
-        $detalle = DetalleIngreso::findOrFail($id);
+        $detalle = DetalleIngreso::find($id);
+
+        if (!$detalle) {
+            return response()->json(['message' => 'Detalle de ingreso no encontrado'], 404);
+        }
+
         return response()->json($detalle, 200);
     }
 
     public function update(Request $request, string $id)
     {
-        $detalle = DetalleIngreso::findOrFail($id);
+        $detalle = DetalleIngreso::find($id);
+        if (!$detalle) {
+            return response()->json(['message' => 'Detalle de ingreso no encontrado'], 404);
+        }
+        $request->validate([
+            'ingreso_id' => 'required|integer|exists:ingresos,id',
+            'producto_id' => 'required|integer|exists:productos,id',
+            'cantidad' => 'required|integer|min:1',
+            'precio_compra' => 'required|numeric',
+        ]);
         $detalle->update($request->all());
         return response()->json($detalle, 200);
     }
 
     public function destroy(string $id)
     {
-        $detalle = DetalleIngreso::findOrFail($id);
+        $detalle = DetalleIngreso::find($id);
+        if (!$detalle) {
+            return response()->json(['message' => 'Detalle de ingreso no encontrado'], 404);
+        }
         $detalle->delete();
         return response()->json(['mensaje' => 'Detalle eliminado correctamente'], 200);
     }
