+++ app/Http/Controllers/Api/ProductoController.php
@@ -17,26 +17,49 @@ public function index()
 
     public function store(Request $request)
     {
+        $request->validate([
+            'categoria_id' => 'required|integer|exists:categorias,id',
+            'nombre' => 'required|string|max:100',
+            'precio' => 'required|numeric',
+            'stock' => 'required|integer',
+        ]);
+
         $producto = Producto::create($request->all());
         return response()->json($producto, 201);
     }
 
     public function show(string $id)
     {
-        $producto = Producto::with('categoria')->findOrFail($id);
+        $producto = Producto::with('categoria')->find($id);
+        if (!$producto) {
+            return response()->json(['message' => 'Producto no encontrado'], 404);
+        }
         return response()->json($producto, 200);
     }
 
     public function update(Request $request, string $id)
     {
-        $producto = Producto::findOrFail($id);
+        $producto = Producto::find($id);
+        if (!$producto) {
+            return response()->json(['message' => 'Producto no encontrado'], 404);
+        }
+        $request->validate([
+            'categoria_id' => 'required|integer|exists:categorias,id',
+            'nombre' => 'required|string|max:100',
+            'precio' => 'required|numeric',
+            'stock' => 'required|integer',
+        ]);
         $producto->update($request->all());
         return response()->json($producto, 200);
     }
 
     public function destroy(string $id)
     {
-        $producto = Producto::findOrFail($id);
+        $producto = Producto::find($id);
+        if (!$producto) {
+            return response()->json(['message' => 'Producto no encontrado'], 404);
+        }
+
         $producto->delete();
         return response()->json(['mensaje' => 'Producto eliminado correctamente'], 200);
     }
